+++ src/microservices/Microservices.IsIdentifiable/Options/IsIdentifiableServiceOptions.cs
@@ -6,7 +6,7 @@ using CommandLine.Text;
 
 namespace Microservices.IsIdentifiable.Options
 {
-    [Verb("service")]
+    [Verb("service", HelpText = "Run tool as a service processing data in a RabbitMQ queue")]
     public class IsIdentifiableServiceOptions : IsIdentifiableAbstractOptions
     {
         [Option('y', HelpText = "Configuration file", Required = true)]